Reworking the Maze. User Experience improvements.
Added a start button to accompany the row/column selector so that the maze starts when the start button is selected. Also inserted a page refresh so that the maze doesn't bonk out when the user tries to run it again. It still needs to capture the size selected when the refresh occurs, that's what I'll implement next.