slight progress in new presentation timing
Now when user clicks start, the walls are generated and displayed. This is currently set to 50ms per wall, may tweak this or allow it to be a user setting later. Next step will be to place the Entry and Exit tiles and pause a moment to allow the user to understand what the entry & exit nodes are before the search begins.